requirements.txt
pyusb
pyyaml
